Move exceptions and create base app exception
