+++ backend/src/main/java/pl/lodz/p/it/opinioncollector/exceptions/BaseApplicationException.java
@@ -1,0 +1,3 @@
+package pl.lodz.p.it.opinioncollector.exceptions;
+
+public abstract class BaseApplicationException extends Exception { }
