+++ backend/src/main/java/pl/lodz/p/it/opinioncollector/exceptions/opinion/OpinionNotFoundException.java
@@ -1,0 +1,5 @@
+package pl.lodz.p.it.opinioncollector.exceptions.opinion;
+
+import pl.lodz.p.it.opinioncollector.exceptions.BaseApplicationException;
+
+public class OpinionNotFoundException extends BaseApplicationException { }
