+++ backend/src/main/java/pl/lodz/p/it/opinioncollector/exceptions/opinion/OpinionOperationAccessForbiddenException.java
@@ -1,0 +1,5 @@
+package pl.lodz.p.it.opinioncollector.exceptions.opinion;
+
+import pl.lodz.p.it.opinioncollector.exceptions.BaseApplicationException;
+
+public class OpinionOperationAccessForbiddenException extends BaseApplicationException { }
