--- backend/src/main/java/pl/lodz/p/it/opinioncollector/opinion/exceptions/OpinionNotFoundException.java
@@ -1,6 +1,0 @@
-package pl.lodz.p.it.opinioncollector.opinion.exceptions;
-
-// FIXME create base application exception
-public class OpinionNotFoundException extends Exception {
-    public OpinionNotFoundException() { }
-}
