--- backend/src/main/java/pl/lodz/p/it/opinioncollector/opinion/exceptions/OpinionOperationAccessForbiddenException.java
@@ -1,5 +1,0 @@
-package pl.lodz.p.it.opinioncollector.opinion.exceptions;
-
-// FIXME create base application exception
-public class OpinionOperationAccessForbiddenException extends Exception {
-}
